All AbstractHealthPoints now treat health as a float instead of a double.

+++ Assets/Scripts/Damage and Health/AbstractHealthPoints.cs
@@ -69,7 +69,7 @@ public abstract class AbstractHealthPoints : MonoBehaviour
         return isDead;
     }
 
-    public abstract void DealDamage(double amount);    //Deals the given amount of damage.  Different implementations can have different ways of handling this.
+    public abstract void DealDamage(float amount);    //Deals the given amount of damage.  Different implementations can have different ways of handling this.
                                                     //Must broadcast the "OnTakeDamage" message.
 
     public virtual void AttackFrom(DamageSource src)
